# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/enums": {
      "get": {
        "operationId": "EnumsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnumCatalogResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Enums"
        ]
      }
    },
    "/enums/{name}": {
      "get": {
        "operationId": "EnumsController_findOne",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnumCatalogResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "그 이름의 enum 이 없다"
          }
        },
        "tags": [
          "Enums"
        ]
      }
    },
    "/staff-members/{id}/retirement-preview": {
      "get": {
        "operationId": "RetirementController_preview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "retiredDate",
            "required": true,
            "in": "query",
            "description": "퇴직일(YYYY-MM-DD, 한국 날짜). 그날까지 재직이고 다음 날 0시에 퇴직이 확정된다. 오늘 이후이거나 최근\n3개월 안의 지난 날짜만 받고, 입사일보다 앞설 수 없다.",
            "schema": {
              "example": "2026-10-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetirementPreviewResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "퇴직일 형식 오류 · 3개월보다 오래됨 · 입사일보다 앞섬"
          },
          "404": {
            "description": "없는 직원이거나 관리 범위 밖"
          },
          "409": {
            "description": "가입 전 · 이미 퇴직 · 이미 퇴직 예정(처리) / 퇴직 예정이 아님 · 퇴직일 당일 이후 · 같은 날짜(취소 · 변경)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "퇴직이 확정될 때 정리될 근무스케줄 · 개인 TO-DO · 대기 계약. 확인창에 보여 준다.",
        "tags": [
          "Retirement"
        ]
      }
    },
    "/staff-members/{id}/retirement": {
      "post": {
        "operationId": "RetirementController_retire",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetirementDateDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "퇴직일 형식 오류 · 3개월보다 오래됨 · 입사일보다 앞섬"
          },
          "404": {
            "description": "없는 직원이거나 관리 범위 밖"
          },
          "409": {
            "description": "가입 전 · 이미 퇴직 · 이미 퇴직 예정(처리) / 퇴직 예정이 아님 · 퇴직일 당일 이후 · 같은 날짜(취소 · 변경)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "퇴직 처리. 퇴직일만 저장하고 재직으로 둔다(퇴직 예정). 계약은 바꾸지 않는다. 퇴직일 다음 날 0시 배치가\n퇴직으로 바꾸며 정리하고, 지난 날짜면 지금 바로 정리한다. 가입 전 · 이미 퇴직 · 이미 퇴직 예정이면 409.",
        "tags": [
          "Retirement"
        ]
      },
      "put": {
        "operationId": "RetirementController_changeDate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetirementDateDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "퇴직일 형식 오류 · 3개월보다 오래됨 · 입사일보다 앞섬"
          },
          "404": {
            "description": "없는 직원이거나 관리 범위 밖"
          },
          "409": {
            "description": "가입 전 · 이미 퇴직 · 이미 퇴직 예정(처리) / 퇴직 예정이 아님 · 퇴직일 당일 이후 · 같은 날짜(취소 · 변경)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "퇴직일 변경. 퇴직 예정이고 퇴직일 전날까지만 된다(그 밖은 409).",
        "tags": [
          "Retirement"
        ]
      },
      "delete": {
        "operationId": "RetirementController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "퇴직일 형식 오류 · 3개월보다 오래됨 · 입사일보다 앞섬"
          },
          "404": {
            "description": "없는 직원이거나 관리 범위 밖"
          },
          "409": {
            "description": "가입 전 · 이미 퇴직 · 이미 퇴직 예정(처리) / 퇴직 예정이 아님 · 퇴직일 당일 이후 · 같은 날짜(취소 · 변경)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "퇴직 취소. 퇴직 예정이고 퇴직일 전날까지만 된다(그 밖은 409).",
        "tags": [
          "Retirement"
        ]
      }
    }
  },
  "info": {
    "title": "Whale ERP API",
    "description": "품목과 재고 이동을 다루는 ERP API",
    "version": "0.0.1",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "EnumValueDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "description": "저장·전송 값. 예: PENDING_SEND"
          },
          "label": {
            "type": "string",
            "description": "화면에 보이는 한글. 예: 발송 대기"
          },
          "order": {
            "type": "number",
            "description": "선택지 순서. 1부터."
          }
        },
        "required": [
          "value",
          "label",
          "order"
        ]
      },
      "EnumCatalogResponseDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "string",
            "description": "enum 내용의 해시. 내용이 바뀌면 바뀐다 — 클라이언트는 이 값이 같으면 캐시를 그대로 쓴다.",
            "example": "37eab9c919dce591"
          },
          "enums": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/EnumValueDto"
              }
            },
            "example": {
              "ContractStatus": [
                {
                  "value": "PENDING_SEND",
                  "label": "발송 대기",
                  "order": 1
                },
                {
                  "value": "PENDING_SIGNATURE",
                  "label": "서명 대기",
                  "order": 2
                },
                {
                  "value": "SIGNED",
                  "label": "체결 완료",
                  "order": 3
                }
              ]
            }
          }
        },
        "required": [
          "version",
          "enums"
        ]
      },
      "ContractStatus": {
        "type": "string",
        "enum": [
          "PENDING_SEND",
          "PENDING_SIGNATURE",
          "SIGNED",
          "REJECTED",
          "EXPIRED",
          "ENDED"
        ],
        "description": "발송 대기 또는 서명 대기"
      },
      "ContractToEndDto": {
        "type": "object",
        "properties": {
          "status": {
            "description": "발송 대기 또는 서명 대기",
            "allOf": [
              {
                "$ref": "#/components/schemas/ContractStatus"
              }
            ]
          },
          "contractId": {
            "type": "number"
          }
        },
        "required": [
          "status",
          "contractId"
        ]
      },
      "RetirementPreviewResponseDto": {
        "type": "object",
        "properties": {
          "schedulesToRemove": {
            "type": "number",
            "description": "퇴직일 다음 날부터의 근무스케줄 수 — 지운다"
          },
          "todosToUnassign": {
            "type": "number",
            "description": "개인으로 배정된 미완료 TO-DO 수 — 배정을 푼다"
          },
          "contractsToEnd": {
            "description": "발송 대기 · 서명 대기 계약 — 종료로 닫는다",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContractToEndDto"
            }
          }
        },
        "required": [
          "schedulesToRemove",
          "todosToUnassign",
          "contractsToEnd"
        ]
      },
      "RetirementDateDto": {
        "type": "object",
        "properties": {
          "retiredDate": {
            "type": "string",
            "description": "퇴직일(YYYY-MM-DD, 한국 날짜). 그날까지 재직이고 다음 날 0시에 퇴직이 확정된다. 오늘 이후이거나 최근\n3개월 안의 지난 날짜만 받고, 입사일보다 앞설 수 없다.",
            "example": "2026-10-31"
          }
        },
        "required": [
          "retiredDate"
        ]
      }
    }
  }
}